Set validation for vacation type
