+++ web/src/main/java/com/infoshareacademy/gitloopersi/validator/VacationDefiningValidator.java
@@ -137,9 +137,8 @@ public class VacationDefiningValidator {
 
     logger.info("Validating vacation type {}", vacationType);
 
-    return vacationType.equals(VacationType.VACATION_LEAVE.getType()) || vacationType
-        .equals(VacationType.SPECIAL_LEAVE.getType()) || vacationType
-        .equals(VacationType.CHILDCARE.getType());
+    return VacationType.VACATION_LEAVE.getType().equals(vacationType) || VacationType.SPECIAL_LEAVE
+        .getType().equals(vacationType) || VacationType.CHILDCARE.getType().equals(vacationType);
   }
 
   public int calculateVacationPoolForEmployee(Long employeeId) throws IOException {
