+++ web/src/main/webapp/WEB-INF/web.xml
@@ -12,8 +12,8 @@
     <class-name>com.mysql.cj.jdbc.MysqlDataSource</class-name>
     <name>java:ds/evc</name>
     <password>rootpasswordgiven</password>
-<!--    <url>jdbc:mysql://mysql/calendar?allowPublicKeyRetrieval=true&amp;useSSL=false&amp;useLegacyDatetimeCode=false&amp;serverTimezone=Europe/Warsaw</url>-->
-    <url>jdbc:mysql://0.0.0.0:3308/calendar?allowPublicKeyRetrieval=true&amp;useSSL=false&amp;useLegacyDatetimeCode=false&amp;serverTimezone=Europe/Warsaw</url>
+    <url>jdbc:mysql://mysql/calendar?allowPublicKeyRetrieval=true&amp;useSSL=false&amp;useLegacyDatetimeCode=false&amp;serverTimezone=Europe/Warsaw</url>
+<!--    <url>jdbc:mysql://0.0.0.0:3308/calendar?allowPublicKeyRetrieval=true&amp;useSSL=false&amp;useLegacyDatetimeCode=false&amp;serverTimezone=Europe/Warsaw</url>-->
     <user>root</user>
   </data-source>
